eval/nuggetizer/requirements.txt
openai>=1.55.3
python-dotenv>=0.19.0
typing-extensions>=4.0.0
requests>=2.25.0
tqdm>=4.65.0 
tiktoken>=0.8